\name{zscorem}
\alias{zscorem}
\title{zscorem: Transformation of monthly m-hydroclimatic variables into m-zscores indexes over 12-months each one}
\usage{
zcorem(data)
}
\description{
Calculation of seasonal sum matrix indexes for a season of n-months from m-hydroclimatic variables
}
\arguments{
data: A matrix object, monthly hydroclimatic variables
}
\examples{
zcorem(data=database3)
}
